requirements.txt
openai>=1.12.0
anthropic>=0.18.0
tiktoken>=0.6.0
python-dotenv>=1.0.0
tenacity>=8.2.0
chromadb>=0.4.22
